requirements.txt
amply==0.1.4
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
dnspython==2.0.0
docutils==0.16
email-validator==1.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jwcrypto==0.8
MarkupSafe==1.1.1
oauth2client==4.1.3
packaging==20.8
pluggy==0.13.1
protobuf==3.14.0
PuLP==2.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pyparsing==2.4.7
Pyrebase4==4.4.3
pytest==6.2.1
python-jwt==3.3.0
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
